Add support for dynamic responses
Resolves #16.

It is now possible to create responses dynamically. This enables greater
customisation on a per-request basis. For example:

```Delphi
WebMock.StubRequest('*', '*')
  .ToRespondWith(
    procedure (const ARequest: IWebMockHTTPRequest;
               const AResponse: IWebMockResponseBuilder)
    begin
      AReponse
        .WithStatus(202)
        .WithHeader('header-1', 'a-value')
        .WithBody('Some content...');
    end
  );
```

The `ARequest: IWebMockHTTPRequest` parameter is available for
inspection.

The internals have been updated to introduce a "responder" layer
allowing different handlers for responses which should be useful beyond
this change.